Move repeatingGroup as custom in EditFormComponent and delete EditFormContainer
