+++ frontend/language/src/nb.json
@@ -1631,8 +1631,6 @@
   "ux_editor.modal_properties_group_add_button_description": "Teksten legger seg på slutten av teksten i knappen. “Legg til ny (din tekst)”",
   "ux_editor.modal_properties_group_add_components": "Legg til skjemakomponent i gruppen",
   "ux_editor.modal_properties_group_change_id": "Gruppe-ID",
-  "ux_editor.modal_properties_group_id_not_unique_error": "Gruppen må ha en unik id",
-  "ux_editor.modal_properties_group_id_not_valid": "Gruppe id kan kun bestå av engelske bokstaver, tall og '-' tegnet.",
   "ux_editor.modal_properties_group_max_occur": "Maksimalt antall repetisjoner",
   "ux_editor.modal_properties_group_repeating": "Repeterende gruppe",
   "ux_editor.modal_properties_group_table_headers": "Kryss av for de titlene som skal være synlige i tabellvisningen",
