+++ frontend/libs/studio-icons/.eslintrc.js
@@ -14,7 +14,7 @@ module.exports = {
               match: true,
             },
           },
-        ]
+        ],
       },
     },
     {
