+++ frontend/libs/studio-icons/src/react/icons/ConfirmationTaskIcon.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import type { IconProps } from '../types';
 import { SvgTemplate } from './SvgTemplate';
 
-export const ConfirmationTaskIcon = (props: IconProps):React.ReactElement => {
+export const ConfirmationTaskIcon = (props: IconProps): React.ReactElement => {
   return (
     <SvgTemplate {...props}>
       <rect
