+++ frontend/libs/studio-icons/src/react/icons/DataTaskIcon.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import type { IconProps } from '../types';
 import { SvgTemplate } from './SvgTemplate';
 
-export const DataTaskIcon = (props: IconProps):React.ReactElement => {
+export const DataTaskIcon = (props: IconProps): React.ReactElement => {
   return (
     <SvgTemplate {...props}>
       <rect
