+++ frontend/packages/ux-editor/src/components/Properties/Properties.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from 'react';
 import { Calculations } from './Calculations';
-import { Content } from './Content';
 import { Text } from './Text';
 import { useTranslation } from 'react-i18next';
 import { Accordion } from '@digdir/design-system-react';
@@ -9,6 +8,7 @@ import classes from './Properties.module.css';
 import { Dynamics } from './Dynamics';
 import { PropertiesHeader } from './PropertiesHeader';
 import { isContainer } from '../../utils/formItemUtils';
+import { EditFormComponent } from '../config/EditFormComponent';
 
 export const Properties = () => {
   const { t } = useTranslation();
@@ -56,7 +56,18 @@ export const Properties = () => {
             {t('right_menu.content')}
           </Accordion.Header>
           <Accordion.Content>
-            {formId ? <Content /> : t('right_menu.content_empty')}
+            {formId ? (
+              <EditFormComponent
+                editFormId={formId}
+                component={form}
+                handleComponentUpdate={async (updatedComponent) => {
+                  handleUpdate(updatedComponent);
+                  debounceSave(formId, updatedComponent);
+                }}
+              />
+            ) : (
+              t('right_menu.content_empty')
+            )}
           </Accordion.Content>
         </Accordion.Item>
         <Accordion.Item open={openList.includes('dynamics')}>
