+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/RepeatingGroup/index.ts
@@ -1,0 +1,1 @@
+export { RepeatingGroupComponent } from './RepeatingGroupComponent';
