+++ frontend/packages/ux-editor/src/testing/componentSchemaMocks.ts
@@ -1,7 +1,6 @@
 import AlertSchema from './schemas/json/component/Alert.schema.v1.json';
-// TODO: Add schemas for the commented out components (https://github.com/Altinn/altinn-studio/issues/10868):
 import AccordionSchema from './schemas/json/component/Accordion.schema.v1.json';
-// import AccordionGroupSchema from './schemas/json/component/AccordionGroup.schema.v1.json';
+import AccordionGroupSchema from './schemas/json/component/AccordionGroup.schema.v1.json';
 import ActionButtonSchema from './schemas/json/component/ActionButton.schema.v1.json';
 import AddressComponentSchema from './schemas/json/component/AddressComponent.schema.v1.json';
 import AttachmentListSchema from './schemas/json/component/AttachmentList.schema.v1.json';
@@ -32,10 +31,12 @@ import PanelSchema from './schemas/json/component/Panel.schema.v1.json';
 import ParagraphSchema from './schemas/json/component/Paragraph.schema.v1.json';
 import PrintButtonSchema from './schemas/json/component/PrintButton.schema.v1.json';
 import RadioButtonsSchema from './schemas/json/component/RadioButtons.schema.v1.json';
+//import RepeatingGroupSchema from './schemas/json/component/RepeatingGroup.schema.v1.json';
 import SummarySchema from './schemas/json/component/Summary.schema.v1.json';
 import TextAreaSchema from './schemas/json/component/TextArea.schema.v1.json';
 
 export const componentSchemaMocks = {
+  AccordionGroup: AccordionGroupSchema,
   Accordion: AccordionSchema,
   ActionButton: ActionButtonSchema,
   AddressComponent: AddressComponentSchema,
@@ -68,6 +69,7 @@ export const componentSchemaMocks = {
   Paragraph: ParagraphSchema,
   PrintButton: PrintButtonSchema,
   RadioButtons: RadioButtonsSchema,
+  //RepeatingGroup: RepeatingGroupSchema,
   Summary: SummarySchema,
   TextArea: TextAreaSchema,
 };
